refactor : Domain 클래스를 리팩토링 한다
