+++ src/main/java/ssu/eatssu/domain/menu/Menu.java
@@ -2,7 +2,6 @@ package ssu.eatssu.domain.menu;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.hibernate.annotations.ColumnDefault;
 import ssu.eatssu.domain.restaurant.Restaurant;
 
 import java.util.ArrayList;
@@ -24,26 +23,8 @@ public class Menu {
 
     private Integer price;
 
-    @ColumnDefault("0.0")
-    private Double mainRate = 0.0;
-
-    @ColumnDefault("0.0")
-    private Double amountRate = 0.0;
-
-    @ColumnDefault("0.0")
-    private Double tasteRate = 0.0;
-
-    @ColumnDefault("0")
-    private Integer totalMainRate = 0;
-
-    @ColumnDefault("0")
-    private Integer totalAmountRate = 0;
-
-    @ColumnDefault("0")
-    private Integer totalTasteRate = 0;
-
-    @ColumnDefault("0")
-    private Integer reviewCount = 0;
+    @Enumerated(EnumType.STRING)
+    private MenuType menuType;
 
     @OneToMany(mappedBy = "menu", cascade = CascadeType.ALL)
     private List<MealMenu> mealMenus = new ArrayList<>();
@@ -55,7 +36,6 @@ public class Menu {
     @Embedded
     private Reviews reviews = new Reviews();
 
-    //정적 팩토리 메서드
     private Menu(String name, Restaurant restaurant, Integer price) {
         this.name = name;
         this.restaurant = restaurant;
@@ -85,55 +65,8 @@ public class Menu {
     public void addReview(Review review) {
         reviews.add(review);
     }
-//
-//    public void addReview(Integer mainRate, Integer tasteRate, Integer amountRate) {
-//        this.reviewCount++;
-//        this.totalMainRate += mainRate;
-//        this.totalTasteRate += tasteRate;
-//        this.totalAmountRate += amountRate;
-//        calculateRate();
-//    }
-//
-//    public void deleteReview() {
-//        refreshReview();
-//    }
-//
-//    public void updateReview() {
-//        refreshReview();
-//    }
-//
-//    public void refreshReview() {
-//        int totalMain = 0;
-//        int totalTaste = 0;
-//        int totalAmount = 0;
-//        for (Review review : this.reviews) {
-//            totalMain += review.getMainRate();
-//            totalTaste += review.getTasteRate();
-//            totalAmount += review.getAmountRate();
-//        }
-//        this.totalMainRate = totalMain;
-//        this.totalTasteRate = totalTaste;
-//        this.totalAmountRate = totalAmount;
-//        this.reviewCount = (int) reviews.stream().count();
-//        calculateRate();
-//    }
-//
-//    private void calculateRate() { // 평점 계산 후 적용
-//        if (this.reviewCount == 0) {
-//            rateReset();
-//        } else {
-//            this.mainRate = this.totalMainRate.doubleValue() / this.reviewCount.doubleValue();
-//            this.tasteRate = this.totalTasteRate.doubleValue() / this.reviewCount.doubleValue();
-//            this.amountRate = this.totalAmountRate.doubleValue() / this.reviewCount.doubleValue();
-//        }
-//    }
-//
-//    private void rateReset() { //평점 초기화
-//        this.totalMainRate = 0;
-//        this.totalTasteRate = 0;
-//        this.totalAmountRate = 0;
-//        this.mainRate = 0.0;
-//        this.tasteRate = 0.0;
-//        this.amountRate = 0.0;
-//    }
+
+    public int getTotalReviewCount() {
+        return reviews.size();
+    }
 }
