+++ src/main/java/ssu/eatssu/web/review/dto/ReviewInformationResponse.java
@@ -1,0 +1,5 @@
+package ssu.eatssu.web.review.dto;
+
+public interface ReviewInformationResponse {
+
+}
